* Added some precision to the action update timing + Added RotateBy/To

+++ GREATClient/BaseClass/Drawable.cs
@@ -51,7 +51,7 @@ namespace GREATClient.BaseClass
 		/// Gets or sets the orientation.
 		/// </summary>
 		/// <value>The orientation.</value>
-		public float Orientation { get; set; }
+		public double Orientation { get; set; }
 
 		/// <summary>
 		/// Gets or sets the source rectangle.
