+++ GREATClient/BaseClass/DrawableCircleContour.cs
@@ -66,7 +66,7 @@ namespace GREATClient.BaseClass
 		protected override void OnDraw(SpriteBatch batch)
 		{
 			batch.Begin();
-			batch.Draw(Texture,GetAbsolutePosition(),SourceRectangle,Tint,Orientation,
+			batch.Draw(Texture,GetAbsolutePosition(),SourceRectangle,Tint,(float)Orientation,
 			           OriginRelative * new Vector2(Texture.Width, Texture.Height),Scale,Effects,0);
 			batch.End();
